feat: smart vault manager

+++ src/interfaces/IUsdPriceFeedManager.sol
@@ -3,6 +3,6 @@ pragma solidity ^0.8.16;
 
 interface IUsdPriceFeedManager {
     function usdDecimals() external view returns (uint256 usdDecimals);
-    function assetToUsd(address asset, uint assetAmount) external view returns (uint256 usdAmount);
-    function usdToAsset(address asset, uint usdAmount) external view returns (uint256 assetAmount);
+    function assetToUsd(address asset, uint256 assetAmount) external view returns (uint256 usdAmount);
+    function usdToAsset(address asset, uint256 usdAmount) external view returns (uint256 assetAmount);
 }
